requirements.txt
Flask==2.3.3
Flask-JWT-Extended>=4.6
Flask-SQLAlchemy>=3.1
Flask-Bcrypt>=1.0
Flask-SocketIO>=5.3
redis>=5.0
python-dotenv>=1.0
gunicorn>=21.0
eventlet>=0.33
# Added for Feature 7 real image processing & Azure storage
Pillow>=10.0.0
azure-storage-blob>=12.19.0
